move `StreamWaker` to utils
